Revert "Added compatibility with Python 3 and some other improvements"
