--- manifest-translated.ini.tpl
@@ -1,2 +1,0 @@
-summary = "{addon_summary}"
-description = """{addon_description}"""
