--- manifest.ini.tpl
@@ -1,10 +1,0 @@
-name = {addon_name}
-summary = "{addon_summary}"
-description = """{addon_description}"""
-author = "{addon_author}"
-url = {addon_url}
-version = {addon_version}
-docFileName = {addon_docFileName}
-minimumNVDAVersion = {addon_minimumNVDAVersion}
-lastTestedNVDAVersion = {addon_lastTestedNVDAVersion}
-updateChannel = {addon_updateChannel}
