--- sconstruct
@@ -1,183 +1,0 @@
-# NVDA add-on template  SCONSTRUCT file
-#Copyright (C) 2012, 2014 Rui Batista <ruiandrebatista@gmail.com>
-#This file is covered by the GNU General Public License.
-#See the file COPYING.txt for more details.
-
-import gettext
-import os
-import os.path
-import zipfile
-import sys
-sys.dont_write_bytecode = True
-
-import buildVars
-
-if sys.version_info.major == 2:
-	import codecs
-	open = codecs.open
-
-
-def md2html(source, dest):
-	import markdown
-	lang = os.path.basename(os.path.dirname(source)).replace('_', '-')
-	localeLang = os.path.basename(os.path.dirname(source))
-	try:
-		_ = gettext.translation("nvda", localedir=os.path.join("addon", "locale"), languages=[localeLang]).ugettext if sys.version_info.major == 2 else gettext.translation("nvda", localedir=os.path.join("addon", "locale"), languages=[localeLang]).gettext
-		title=u"{0}".format(_(buildVars.addon_info["addon_summary"]))
-	except:
-		title="{0}".format(buildVars.addon_info["addon_summary"]) 
-	headerDic = {
-		"[[!meta title=\"": "# ",
-		"\"]]": " #",
-	}
-	with open(source, "r", encoding = "utf-8") as f:
-		mdText = f.read()
-		headerList = headerDic.iteritems () if sys.version_info.major == 2 else list(headerDic.items())
-		for k, v in headerList:
-			mdText = mdText.replace(k, v, 1)
-		htmlText = markdown.markdown(mdText)
-	with open(dest, "w", encoding = "utf-8") as f:
-		f.write("<!DOCTYPE html>\n" +
-			"<html lang=\"%s\">\n" % lang +
-			"<head>\n" +
-			"<meta charset=\"UTF-8\">\n" +
-			"<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\n" +
-			"<link rel=\"stylesheet\" type=\"text/css\" href=\"../style.css\" media=\"screen\">\n" +
-			"<title>%s</title>\n" % title +
-			"</head>\n<body>\n"
-		)
-		f.write(htmlText)
-		f.write("\n</body>\n</html>")
-
-def generateHelpFiles (source, target, env, for_signature):
-	action = env.Action(lambda target, source, env : md2html(source[0].abspath, target[0].abspath) and None,
-	lambda target, source, env : "Generating %s" % target[0])
-	return action
-
-env = Environment(ENV=os.environ, tools=['gettexttool'])
-env.Append(**buildVars.addon_info)
-
-addonFile = env.File("${addon_name}-${addon_version}.nvda-addon")
-
-def addonGenerator(target, source, env, for_signature):
-	action = env.Action(lambda target, source, env : createAddonBundleFromPath(source[0].abspath, target[0].abspath) and None,
-	lambda target, source, env : "Generating Addon %s" % target[0])
-	return action
-
-def manifestGenerator(target, source, env, for_signature):
-	action = env.Action(lambda target, source, env : generateManifest(source[0].abspath, target[0].abspath) and None,
-	lambda target, source, env : "Generating manifest %s" % target[0])
-	return action
-
-
-def translatedManifestGenerator(target, source, env, for_signature):
-	dir = os.path.abspath(os.path.join(os.path.dirname(str(source[0])), ".."))
-	lang = os.path.basename(dir)
-	action = env.Action(lambda target, source, env : generateTranslatedManifest(source[1].abspath, lang, target[0].abspath) and None,
-	lambda target, source, env : "Generating translated manifest %s" % target[0])
-	return action
-
-env['BUILDERS']['NVDAAddon'] = Builder(generator=addonGenerator)
-env['BUILDERS']['markdown']=Builder(generator = generateHelpFiles,
-	suffix='.html',
-	src_suffix='.md')
-env['BUILDERS']['NVDAManifest'] = Builder(generator=manifestGenerator)
-env['BUILDERS']['NVDATranslatedManifest'] = Builder(generator=translatedManifestGenerator)
-
-def createAddonHelp(dir):
-	docsDir = os.path.join(dir, "doc")
-	if os.path.isfile("style.css"):
-		cssPath = os.path.join(docsDir, "style.css")
-		cssTarget = env.Command(cssPath, "style.css", Copy("$TARGET", "$SOURCE"))
-		env.Depends(addon, cssTarget)
-	if os.path.isfile("readme.md"):
-		readmePath = os.path.join(docsDir, "en", "readme.md")
-		readmeTarget = env.Command(readmePath, "readme.md", Copy("$TARGET", "$SOURCE"))
-		env.Depends(addon, readmeTarget)
-
-
-
-def createAddonBundleFromPath(path, dest):
-	""" Creates a bundle from a directory that contains an addon manifest file."""
-	basedir = os.path.abspath(path)
-	with zipfile.ZipFile(dest, 'w', zipfile.ZIP_DEFLATED) as z:
-		# FIXME: the include/exclude feature may or may not be useful. Also python files can be pre-compiled.
-		for dir, dirnames, filenames in os.walk(basedir):
-			relativePath = os.path.relpath(dir, basedir)
-			for filename in filenames:
-				pathInBundle = os.path.join(relativePath, filename)
-				absPath = os.path.join(dir, filename)
-				if pathInBundle not in buildVars.excludedFiles: z.write(absPath, pathInBundle)
-	return dest
-
-def generateManifest(source, dest):
-	with open(source, "r", encoding = "utf-8") as f:
-		manifest_template = f.read()
-	manifest = manifest_template.format(**buildVars.addon_info)
-	with open(dest, "w", encoding = "utf-8") as f:
-		f.write(manifest)
-
-def generateTranslatedManifest(source, language, out):
-	_ = gettext.translation("nvda", localedir=os.path.join("addon", "locale"), languages=[language]).ugettext if sys.version_info.major == 2 else gettext.translation("nvda", localedir=os.path.join("addon", "locale"), languages=[language]).gettext
-	vars = {}
-	for var in ("addon_summary", "addon_description"):
-		if isinstance(buildVars.addon_info[var], str):
-			vars[var] = _(buildVars.addon_info[var])
-		elif isinstance(buildVars.addon_info[var], list):
-			vars[var] = ''.join([_(l) for l in buildVars.addon_info[var]])
-		else:
-			raise TypeError("Error with %s key in buildVars" % var)
-	with open(source, "r", encoding = "utf-8") as f:
-		manifest_template = f.read()
-	result = manifest_template.format(**vars)
-	with open(out, "w", encoding = "utf-8") as f:
-		f.write(result)
-
-def expandGlobs(files):
-	return [f for pattern in files for f in env.Glob(pattern)]
-
-addon = env.NVDAAddon(addonFile, env.Dir('addon'))
-
-pythonFiles = expandGlobs(buildVars.pythonSources)
-for file in pythonFiles:
-	env.Depends(addon, file)
-
-# Pot target
-i18nFiles = expandGlobs(buildVars.i18nSources)
-gettextvars={
-		'gettext_package_bugs_address' : 'nvda-translations@freelists.org',
-		'gettext_package_name' : buildVars.addon_info['addon_name'],
-		'gettext_package_version' : buildVars.addon_info['addon_version']
-	}
-
-pot = env.gettextPotFile("${addon_name}.pot", i18nFiles, **gettextvars)
-env.Alias('pot', pot)
-env.Depends(pot, i18nFiles)
-mergePot = env.gettextMergePotFile("${addon_name}-merge.pot", i18nFiles, **gettextvars)
-env.Alias('mergePot', mergePot)
-env.Depends(mergePot, i18nFiles)
-
-# Generate Manifest path
-manifest = env.NVDAManifest(os.path.join("addon", "manifest.ini"), os.path.join("manifest.ini.tpl"))
-# Ensure manifest is rebuilt if buildVars is updated.
-env.Depends(manifest, "buildVars.py")
-
-env.Depends(addon, manifest)
-createAddonHelp("addon") # We need at least doc in English and should enable the Help button for the add-on in Add-ons Manager
-langDirs = [f for f in env.Glob(os.path.join("addon", "locale", "*"))]
-
-#Allow all NVDA's gettext po files to be compiled in source/locale, and manifest files to be generated
-for dir in langDirs:
-	poFile = dir.File(os.path.join("LC_MESSAGES", "nvda.po"))
-	moFile=env.gettextMoFile(poFile)
-	env.Depends(moFile, poFile)
-	translatedManifest = env.NVDATranslatedManifest(dir.File("manifest.ini"), [moFile, os.path.join("manifest-translated.ini.tpl")])
-	env.Depends(translatedManifest, ["buildVars.py"])
-	env.Depends(addon, [translatedManifest, moFile])
-	#Convert markdown files to html
-	for mdFile in env.Glob(os.path.join('addon', 'doc', '*', '*.md')):
-		htmlFile = env.markdown(mdFile)
-		env.Depends(htmlFile, [mdFile, moFile])
-		env.Depends(addon, htmlFile)
-env.Default(addon)
-env.Clean (addon, ['.sconsign.dblite', 'addon/doc/en/', 'addon/doc'])
