+++ speechHistory/manifest.ini
@@ -1,0 +1,7 @@
+﻿name = "speechHistory"
+summary = "Speech history review and copying"
+description = "This add-on allows you to review the 100 most recent strings spoken by NVDA, by default using Shift+F11 and Shift+F12.  Additionally, you can copy any spoken item to the clipboard by pressing F12."
+version = "2019.03.30"
+url = "https://github.com/jscholes/nvda-speech-history"
+author = "Tyler Spivey, James Scholes"
+lastTestedNVDAVersion = "2019.1"
